Extract image panel from update image to share with container app creation
